sonarqube issues resolved

+++ app/src/main/java/a75f/io/renatus/views/MasterControl/MasterControl.java
@@ -28,7 +28,6 @@ import a75f.io.renatus.R;
 
 import static a75f.io.logic.bo.util.UnitUtils.fahrenheitToCelsius;
 import static a75f.io.renatus.util.BitmapUtil.getBitmapFromVectorDrawable;
-import static a75f.io.renatus.views.MasterControl.MasterControlView.getTuner;
 
 import java.util.HashMap;
 
@@ -156,14 +155,14 @@ public class MasterControl extends View {
 
         //The 2 and 1.5 are used to slide the number on the bitmap image.
         //Text centered left to right and 1/3 the way down the icon.
-        if( (double) MasterControlView.getTuner(useCelsius.get("id").toString())== TunerConstants.USE_CELSIUS_FLAG_ENABLED) {
+        if(MasterControlView.getTuner(useCelsius.get("id").toString())== TunerConstants.USE_CELSIUS_FLAG_ENABLED) {
             canvas.drawText(String.valueOf(Math.round(temps[stateReflected.ordinal()]) + "\u00B0F  (" + fahrenheitToCelsius(Math.round(temps[stateReflected.ordinal()])) + "\u00B0C)"),
-                    xPos + bitmaps[stateReflected.ordinal()].getWidth() / 2,
-                    (float) (yPos + bitmaps[stateReflected.ordinal()].getHeight() / 2), mTempIconPaint);
+                    xPos + (float) bitmaps[stateReflected.ordinal()].getWidth() / 2,
+                    (float) (yPos + (float) bitmaps[stateReflected.ordinal()].getHeight() / 2), mTempIconPaint);
         } else {
             canvas.drawText(String.valueOf(Math.round(temps[stateReflected.ordinal()])),
                     xPos + bitmaps[stateReflected.ordinal()].getWidth() / 2,
-                    (float) (yPos + bitmaps[stateReflected.ordinal()].getHeight() / 2), mTempIconPaint);
+                    (float) (yPos + (float) bitmaps[stateReflected.ordinal()].getHeight() / 2), mTempIconPaint);
         }
     }
 
@@ -301,7 +300,7 @@ public class MasterControl extends View {
         mArrowHeadLeftBitmap = getBitmapFromVectorDrawable(getContext(), R.drawable.ic_arrowhead_left);
         mArrowHeadRightBitmap = getBitmapFromVectorDrawable(getContext(), R.drawable.ic_arrowhead_right);
 
-        if( (double) MasterControlView.getTuner(useCelsius.get("id").toString())== TunerConstants.USE_CELSIUS_FLAG_ENABLED) {
+        if(MasterControlView.getTuner(useCelsius.get("id").toString())== TunerConstants.USE_CELSIUS_FLAG_ENABLED) {
 
             bitmaps[MasterControlState.UPPER_BUILDING_LIMIT.ordinal()] = bitmaps[MasterControlState.LOWER_BUILDING_LIMIT.ordinal()] =
                     getBitmapFromVectorDrawable(getContext(), R.drawable.black_teardrop_small);
@@ -761,10 +760,10 @@ public class MasterControl extends View {
 
             mDebugTextPaint.setTextSize(mArrowTextSize);
             mDebugTextPaint.getTextBounds(temp, 0, temp.length(), bounds);
-            if( (double) MasterControlView.getTuner(useCelsius.get("id").toString())== TunerConstants.USE_CELSIUS_FLAG_ENABLED) {
-                canvas.drawText(temp + "\u00B0F ("+fahrenheitToCelsius((Double.valueOf(temp)))+"\u00B0C )", getPXForTemp(i) - mSetBack, getTempLineYLocation() + mEnergySavingsSpacing, mDebugTextPaint);
+            if(MasterControlView.getTuner(useCelsius.get("id").toString())== TunerConstants.USE_CELSIUS_FLAG_ENABLED) {
+                canvas.drawText(temp + "\u00B0F ("+fahrenheitToCelsius((Double.valueOf(temp)))+"\u00B0C )", getPXForTemp(i) - mSetBack, getTempLineYLocation() + (float) mEnergySavingsSpacing, mDebugTextPaint);
             } else {
-                canvas.drawText(temp+ " \u00B0F" , getPXForTemp(i) - mSetBack, getTempLineYLocation() + mEnergySavingsSpacing, mDebugTextPaint);
+                canvas.drawText(temp+ " \u00B0F" , getPXForTemp(i) - mSetBack, getTempLineYLocation() + (float) mEnergySavingsSpacing, mDebugTextPaint);
 
             }
         }
